Windows tests work.
