+++ src/test/mod.rs
@@ -1,0 +1,7 @@
+pub mod shared;
+
+#[cfg(unix)]
+pub mod unix;
+
+#[cfg(windows)]
+pub mod windows;
